remove static cast

+++ source/bits.cpp
@@ -3,8 +3,7 @@
 namespace lox {
 
 auto encode_bits(int32_t number) -> std::tuple<uint8_t, uint8_t, uint8_t> {
-  return {static_cast<uint8_t>(number >> 16), static_cast<uint8_t>(number >> 8),
-          static_cast<uint8_t>(number)};
+  return {number >> 16, number >> 8, number};
 }
 
 auto decode_bits(uint8_t a, uint8_t b, uint8_t c) -> int32_t {
